chore(restructure): split subtitles -> overlays/subtitles_{generate,render}.py
Phase 1 step 4: separate concerns. Generation (faster-whisper + SRT
writing) lives in subtitles_generate.py; FFmpeg burn-in + ASS color
conversion lives in subtitles_render.py. Shim at the old path preserves
existing imports. Tests stay 62/62 green.
