+++ openshorts/overlays/subtitles_render.py
@@ -1,0 +1,93 @@
+"""Subtitle burn-in: FFmpeg subtitles filter + ASS color/style conversion."""
+
+import subprocess
+
+
+def hex_to_ass_color(hex_color, opacity=1.0):
+    """Convert #RRGGBB to ASS &HAABBGGRR format. opacity: 0.0=transparent, 1.0=opaque"""
+    hex_color = hex_color.lstrip('#')
+    if len(hex_color) != 6:
+        hex_color = "FFFFFF"
+    r = int(hex_color[0:2], 16)
+    g = int(hex_color[2:4], 16)
+    b = int(hex_color[4:6], 16)
+    alpha = round((1.0 - opacity) * 255)
+    return f"&H{alpha:02X}{b:02X}{g:02X}{r:02X}"
+
+
+def burn_subtitles(video_path, srt_path, output_path, alignment=2, fontsize=16,
+                   font_name="Verdana", font_color="#FFFFFF",
+                   border_color="#000000", border_width=2,
+                   bg_color="#000000", bg_opacity=0.0):
+    """
+    Burns subtitles into the video using FFmpeg.
+    Supports two modes:
+    - Outline mode (bg_opacity=0): Text with colored outline/border
+    - Box mode (bg_opacity>0): Text with semi-transparent background box
+    """
+    # Position mapping
+    ass_alignment = 2
+    align_lower = str(alignment).lower()
+    if align_lower == 'top':
+        ass_alignment = 6
+    elif align_lower == 'middle':
+        ass_alignment = 10
+    elif align_lower == 'bottom':
+        ass_alignment = 2
+
+    # Font size scaling for ASS virtual resolution (PlayResY=288 default)
+    # For vertical 1080x1920 video, we need larger text for readability
+    final_fontsize = int(fontsize * 0.85)
+    if final_fontsize < 10:
+        final_fontsize = 10
+
+    # Path handling for FFmpeg filter syntax
+    safe_srt_path = srt_path.replace('\\', '/').replace(':', '\\:')
+
+    # Convert colors to ASS format and build style
+    primary_colour = hex_to_ass_color(font_color, 1.0)
+
+    if bg_opacity > 0:
+        # Box mode: opaque background box
+        border_style = 3
+        outline_colour = hex_to_ass_color(bg_color, bg_opacity)
+        outline_width = 1
+    else:
+        # Outline mode: text border/outline
+        border_style = 1
+        outline_colour = hex_to_ass_color(border_color, 1.0)
+        outline_width = max(1, border_width)
+
+    back_colour = hex_to_ass_color("#000000", 0.0)
+
+    style_string = (
+        f"Alignment={ass_alignment},"
+        f"Fontname={font_name},"
+        f"Fontsize={final_fontsize},"
+        f"PrimaryColour={primary_colour},"
+        f"OutlineColour={outline_colour},"
+        f"BackColour={back_colour},"
+        f"BorderStyle={border_style},"
+        f"Outline={outline_width},"
+        f"Shadow=0,"
+        f"MarginV=25,"
+        f"Bold=1"
+    )
+
+    cmd = [
+        'ffmpeg', '-y',
+        '-i', video_path,
+        '-vf', f"subtitles='{safe_srt_path}':force_style='{style_string}'",
+        '-c:a', 'copy',
+        '-c:v', 'libx264', '-preset', 'fast', '-crf', '23',
+        output_path
+    ]
+
+    print(f"🎬 Burning subtitles: {' '.join(cmd)}")
+    result = subprocess.run(cmd, stdout=subprocess.DEVNULL, stderr=subprocess.PIPE)
+
+    if result.returncode != 0:
+        print(f"❌ FFmpeg Subtitle Error: {result.stderr.decode()}")
+        raise Exception(f"FFmpeg failed: {result.stderr.decode()}")
+
+    return True
